1. Speed 조절 기능 구현 2. gameplayManager의 역할과 stagePlayer, roundplayer 역할을 분기함 3. mission 팝업 코드 리뷰 및 수정 4. 타워 갯수 변수 타입 변경(float -> int)

+++ Assets/Scripts/Core/XState.cs
@@ -24,8 +24,8 @@ public class XState : States, IState
         }
     }
 
-    float m_TowerCount;
-    public float towerCount
+    int m_TowerCount;
+    public int towerCount
     {
         get => m_TowerCount;
         set { m_TowerCount = value; Set(nameof(towerCount), value); }
@@ -88,7 +88,6 @@ public class XState : States, IState
         get => m_AudioName;
         set { m_AudioName = value; Set(nameof(audioName), value); }
     }
-
 
     void SlientSet()
     {
